Implement AccessControl.for and associated query
This requires supplying a custom query,
`Hyrax::Acl::CustomQueries::FindAccessControl`. For some reason, I
couldn’t get specs to pass when the signature of this query used
`resource:` (with colon), which is what Hyrax’s version of this
query uses. It works fine if you implement it without the colon.

I feel a little weird about the overall architecture of this, since
it expects `Hyrax.query_service` to be defined, and that feels weird
(although manageably so) for applications using this gem on its own.
We may want to consider that integration point in more detail.

(The query service *can* be specified as something other than
`Hyrax.query_service`, which is how the tests handle things, but I’m
not sure this will be ergonomic at actual point of use. Maybe it will
be fine?)